Add shutdown command and fix error crashes.
This PR will add a shutdown command usable by the operator of the bot. It also prevents errors from crashing the bot, instead writing errors to a file and notifying the operator to check the log and file an issue so that they (or someone like me) can fix it.